Minor fixes, rename IconHelper to FileTypeIconHelper
