+++ packages/file-type-icons/README.md
@@ -1,44 +1,35 @@
-# [Office UI Fabric - Icons](http://dev.office.com/fabric)
+# [Office UI Fabric - File Type Icons](http://dev.office.com/fabric)
 
 ##### The React-based front-end framework for building experiences for Office and Office 365.
 
-Fabric Icons includes a collection of 1100+ icons which you can use in your application.
+Fabric File Type Icons includes a collection of icons to represent file types.
 
 
 # Getting started
 
-If you are using Fabric React components, you can make all icons available by calling the `initializeIcons` function from the `@uifabric/icons` package:
+If you are using Fabric React components, you can make all file type icons available by calling the `initializeFileTypeIcons` function from the `@uifabric/file-type-icons` package:
 
 ```tsx
-import { initializeIcons } from '@uifabric/icons';
+import { initializeFileTypeIcons } from '@uifabric/file-type-icons';
 
 // Register icons and pull the fonts from the default SharePoint cdn.
-initializeIcons();
+initializeFileTypeIcons();
 
 // ...or, register icons and pull the fonts from your own cdn:
-initializeIcons('https://my.cdn.com/path/to/icons/');
+initializeFileTypeIcons('https://my.cdn.com/path/to/icons/');
 ```
 
-This will make ALL icons in the collection available, but will download them on demand when referenced using the `@uifabric/styling` APIs `getIcon` or `getIconClassName`.
+This will make ALL icons in the collection available, but will download them on demand.
 
 # Usage in code
 
-## getIcon API
-
-If you are using Fabric React, you can use the `Icon` component and pass in the corresponding iconName property to render a given icon.
+If you are using Fabric React, you can use the `Icon` component and pass in the corresponding icon properties to render a given icon.
 
 ```tsx
 import { Icon } from 'office-ui-fabric-react/lib/Icon';
+import { getFileTypeIconProps } from '@uifabric/file-type-icons/lib/FileTypeIconHelper';
 
-<Icon iconName='Snow' />
-```
-
-The styling package includes a `getIconClassName` api which can provide a css class to use for rendering the icon manually using the `:before` pseudoselector:
-
-```ts
-import { getIconClassName } from '@uifabric/styling';
-
-return `<i class="${getIconClassName('Snow')}" />`;
+<Icon {...getFileTypeIconProps({extension: 'docx', size: '16'}) />
 ```
 
 # Notes
