+++ packages/file-type-icons/package.json
@@ -20,11 +20,11 @@
     "jest"
   ],
   "devDependencies": {
-    "@types/react": "15.0.38",
+    "@types/react": "15.6.7",
     "office-ui-fabric-react-tslint": ">=5.0.0 <6.0.0"
   },
   "dependencies": {
-    "@uifabric/styling": ">=5.10.0 <6.0.0",
+    "@uifabric/styling": ">=5.12.0 <6.0.0",
     "tslib": "^1.6.0"
   }
 }
