+++ packages/file-type-icons/src/FileTypeIconHelper.ts
